Move linear regression animations to new directory location
