+++ animations/linear_regression/data_anim.py
@@ -1,0 +1,9 @@
+from manimlib.imports import *
+
+from lr_utils import *
+
+class Animation(ModelScene):
+    def construct(self):
+        self.custom_setup()
+        self.play(ShowCreation(self.axes, run_time=2, lag_ratio=0.1))
+        self.play(Write(self.data_text), FadeIn(self.dots), lag_ratio=0.1, run_time=4)
