+++ animations/linear_regression/true_function_anim.py
@@ -1,0 +1,16 @@
+from manimlib.imports import *
+
+from lr_utils import *
+
+class Animation(ModelScene):
+    def construct(self):
+        # Generate the axes
+        self.custom_setup()
+        self.add(self.axes, self.data_text, self.dots)
+
+        # Draw Function
+        self.play(Write(self.function_text))
+        self.add(self.function, self.dots)
+        self.play(ShowCreation(self.function))
+        self.wait(duration=1.5)
+
