--- resources/linear_regression/model.py
@@ -1,214 +1,0 @@
-import numpy as np
-
-from manimlib.imports import *
-
-from config import *
-
-class Data(ModelScene):
-    def construct(self):
-        self.custom_setup()
-        self.play(ShowCreation(self.axes, run_time=2, lag_ratio=0.1))
-        self.play(Write(self.data_text), FadeIn(self.dots), lag_ratio=0.1, run_time=4)
-
-
-class TrueFunction(ModelScene):
-    def construct(self):
-        # Generate the axes
-        self.custom_setup()
-        self.add(self.axes, self.data_text, self.dots)
-
-        # Draw Function
-        self.play(Write(self.function_text))
-        self.add(self.function, self.dots)
-        self.play(ShowCreation(self.function))
-        self.wait(duration=1.5)
-
-
-
-class Model(ModelScene):
-    def construct(self):
-        TEXT_SCALE = 0.7
-        TEXT_LBUFF = 0.4
-
-        # Generate the axes
-        self.custom_setup()
-        self.add(self.graph, self.text_group, self.dots)
-
-        # Write Model
-        model_text = TextMobject(r'Model: $y_i$ $=$ {$f($}{$x_i$}{$)$} + $\varepsilon_i$',
-                                 tex_to_color_map={
-                                     '{$f($}': GREEN, '{$)$}': GREEN,
-                                     '$y_i$': BLUE,
-                                     r'$\varepsilon_i$': MAROON
-                                 },
-                                 color=DRAW_COLOR)
-        model_text.to_corner(UP + RIGHT)
-        self.play(Write(model_text))
-
-        # Pick one point to highlight
-        index = 19
-        origin = self.axes.coords_to_point(0, 0, 0)
-        point = self.axes.coords_to_point(XS[index], YS[index], 0)
-        expected_value = self.axes.coords_to_point(XS[index], f(XS[index]), 0)
-
-        # Draw a vertical line from its x to the expected value
-        x_mark = Line(start=[point[0], origin[1], 0],
-                      end=[point[0], origin[1] + .1, 0],
-                      color=DRAW_COLOR)
-        x_text = TexMobject(r'x_7', color=DRAW_COLOR)
-        x_text.next_to(x_mark, direction=DOWN)
-        x_text.scale(TEXT_SCALE)
-
-        dot = Dot(expected_value, color=GREEN)
-        ev_line = DashedLine(start=[point[0], origin[1], 0],
-                             end=[point[0], expected_value[1], 0], color=GREEN)
-
-        self.play(
-            AnimationGroup(FadeIn(x_mark), FadeIn(x_text)),
-            FadeIn(ev_line, duration=2),
-            FadeIn(dot),
-            lag_ration=0.9, duration=4)
-        self.wait()
-
-        ev_text = TextMobject(r'{$f($}{$x_i$}{$)$}',
-                              tex_to_color_map={
-                                     '{$f($}': GREEN, '{$)$}': GREEN,
-                                 }, color=DRAW_COLOR)
-        self.play(ApplyMethod(ev_line.next_to, origin - np.array([0.05, 0, 0]), UP, {'buff': 0}))
-        ev_text.next_to(ev_line, direction=LEFT * TEXT_LBUFF)
-        ev_text.scale(TEXT_SCALE)
-        self.play(FadeIn(ev_text))
-
-
-        # Draw the epislon from the expected to observed
-        epsilon = DashedLine(start=expected_value,
-                             end=point,
-                             color=MAROON)
-        epsilon_text = TexMobject(r'\epsilon_7', color=MAROON)
-        epsilon_text.scale(TEXT_SCALE)
-        epsilon_text.next_to(epsilon, direction=LEFT * TEXT_LBUFF)
-        self.play(ShowCreation(epsilon, duration=1), Write(epsilon_text))
-        self.wait()
-
-        # Draw a line to show how it computes the y from that
-        y_line = DashedLine(start=[point[0], origin[1], 0],
-                            end=[point[0], point[1], 0],
-                            color=BLUE)
-        y_text = TexMobject(r'y_7', color=BLUE)
-        y_text.next_to(y_line, direction=LEFT * TEXT_LBUFF)
-        y_text.scale(TEXT_SCALE)
-        self.play(FadeIn(y_line, duration=2), Write(y_text), lag_ratio=0.5)
-        self.wait(duration=4)
-
-
-class Predictor(LinearScene):
-    def construct(self):
-        self.custom_setup(line_color=BLUE)
-        self.add(self.graph, self.text_group, self.dots)
-
-        # Draw the line and the text
-        self.play(Write(self.predictor_text))
-        self.wait(1)
-        self.play(ShowCreation(self.line))
-        self.wait(2)
-
-        # Highlight error for one point
-        residuals = Group()
-        for x, y in zip(XS, YS):
-            # index = 3 # 10
-            point = self.axes.coords_to_point(x, y, 0)
-            prediction = self.w_0 + self.w_1 * x
-            prediction = self.axes.coords_to_point(x, prediction, 0)
-            residual = DashedLine(start=prediction, end=point, color=RED)
-            residuals.add(residual)
-
-        # Text for residual
-        residual_text = TextMobject(r'Errors observed', color=RED)
-        residual_text.next_to(self.text_group, RIGHT, buff=2)
-
-        # Animate text and residual
-        self.play(Write(residual_text))
-        self.wait()
-        self.add(residuals, self.function, self.line, self.dots) # Order matters
-        self.play(ShowCreation(residuals), duration=4)
-        self.wait(3)
-
-
-class InterpretCoefficients(LinearScene):
-    def construct(self):
-        TEXT_SCALE  = 0.7
-
-        self.custom_setup(line_color=GREEN, graph_shift=[0, 0, 0])
-        self.add(self.axes, self.dots)
-
-        # Model text
-        model_text = TextMobject(r'Model: {$y_i = w_0 + w_1x_i + \varepsilon_i$}',
-                                 color=GREEN,
-                                 tex_to_color_map={r'Model:': DRAW_COLOR})
-        model_text.to_corner(UP + LEFT)
-
-        self.play(Write(model_text))
-        self.add(self.line, self.dots) # To draw below
-        self.play(ShowCreation(self.line))
-
-        # Highlight w_0
-        w_0_point = self.axes.coords_to_point(0, self.w_0, 0)
-        w_0_point = Dot(w_0_point, color=GREEN)
-        w_0_text = TexMobject(r'w_0', color=GREEN)
-        w_0_text.scale(TEXT_SCALE)
-        w_0_text.next_to(w_0_point, LEFT)
-        self.play(Write(w_0_text), FadeIn(w_0_point))
-
-        # Highlight slope
-        def eval(x):
-            return self.w_0 + self.w_1 * x
-
-        start = 1.15
-        end = 2.62 # 1.9
-        start_point = self.axes.coords_to_point(start, eval(start), 0)
-        end_point = self.axes.coords_to_point(end, eval(end), 0)
-
-        corner = self.axes.coords_to_point(end, eval(start), 0)
-
-        # Create run
-        run = DashedLine(start=start_point, end=corner,
-                         color=BLUE)
-        run_text = TexMobject(r'1', color=BLUE)
-        run_text.scale(TEXT_SCALE)
-        run_text.next_to(run, DOWN)
-        self.play(ShowCreation(run), FadeIn(run_text))
-
-        # Create rise
-        rise = DashedLine(start=corner, end=end_point,
-                         color=BLUE)
-        rise_text = TexMobject(r'w_1', color=BLUE)
-        rise_text.scale(TEXT_SCALE)
-        rise_text.next_to(rise, RIGHT, buff=0.05)
-        self.play(ShowCreation(rise), FadeIn(rise_text))
-
-
-        self.wait(3)
-
-class ManyLines(LinearScene):
-    def construct(self):
-        self.custom_setup(line_color=BLUE, graph_shift=[0, 0, 0])
-        self.add(self.axes, self.dots)
-
-        coeffs = [
-            (2, 0.5),
-            (3, 0.1),
-            (1, 0.4),
-            (2.8, -0.02),
-            (3.5, -0.5),
-            (1.5, 0.2)
-        ]
-
-        lines = Group()
-        for w_0, w_1 in coeffs:
-            line = self.linear_function(w_0, w_1, line_color=BLUE)
-            lines.add(line)
-
-        self.add(lines, self.dots)
-        self.play(ShowCreation(lines, run_time=5, lag_ratio=1))
-        self.wait(2)
-
